(ml5717) Renamed rustcoalescence-algorithms-monolithic to rustcoalescence-algorithms-gillespie
