+++ .vscode/settings.json
@@ -5,7 +5,7 @@
     "rust-analyzer.checkOnSave.command": "reap-clippy",
     "rust-analyzer.cargo.allFeatures": false,
     "rust-analyzer.cargo.features": [
-        "rustcoalescence-algorithms-monolithic",
+        "rustcoalescence-algorithms-gillespie",
         "rustcoalescence-algorithms-independent",
         "rustcoalescence-algorithms-cuda",
         "necsim-partitioning-mpi"
