+++ Cargo.toml
@@ -25,7 +25,7 @@ members = [
     "rustcoalescence/scenarios",
 
     "rustcoalescence/algorithms",
-    "rustcoalescence/algorithms/monolithic",
+    "rustcoalescence/algorithms/gillespie",
     "rustcoalescence/algorithms/independent",
     "rustcoalescence/algorithms/cuda",
     "rustcoalescence/algorithms/cuda/kernel",
