+++ necsim/impls/std/Cargo.toml
@@ -14,8 +14,6 @@ necsim-impls-no-std = { path = "../../impls/no-std" }
 
 thiserror = "1.0"
 contracts = { git = "https://gitlab.com/MomoLangenstein/contracts", rev = "3a44375a" }
-keyed_priority_queue = "0.4"
-fxhash = "0.2"
 rand_core = "0.6"
 anyhow = "1.0"
 bincode = "1.3"
