--- necsim/impls/std/src/cogs/active_lineage_sampler/mod.rs
@@ -1,1 +1,0 @@
-pub mod gillespie;
