+++ necsim/impls/std/src/cogs/mod.rs
@@ -1,3 +1,2 @@
-pub mod active_lineage_sampler;
 pub mod dispersal_sampler;
 pub mod rng;
