+++ rustcoalescence/Cargo.toml
@@ -24,7 +24,7 @@ rustcoalescence-algorithms = { path = "algorithms" }
 
 necsim-partitioning-mpi = { path = "../necsim/partitioning/mpi", optional = true }
 
-rustcoalescence-algorithms-monolithic = { path = "algorithms/monolithic", optional = true }
+rustcoalescence-algorithms-gillespie = { path = "algorithms/gillespie", optional = true }
 rustcoalescence-algorithms-independent = { path = "algorithms/independent", optional = true }
 rustcoalescence-algorithms-cuda = { path = "algorithms/cuda", optional = true }
