+++ rustcoalescence/algorithms/gillespie/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rustcoalescence-algorithms-monolithic"
+name = "rustcoalescence-algorithms-gillespie"
 version = "0.1.0"
 authors = ["Momo Langenstein <momo.langenstein17@imperial.ac.uk>"]
 license = "MIT OR Apache-2.0"
