+++ rustcoalescence/algorithms/gillespie/src/classical.rs
@@ -10,7 +10,10 @@ use necsim_core_maths::IntrinsicsMathsCore;
 
 use necsim_impls_no_std::{
     cogs::{
-        active_lineage_sampler::classical::ClassicalActiveLineageSampler,
+        active_lineage_sampler::{
+            alias::individual::IndividualAliasActiveLineageSampler,
+            classical::ClassicalActiveLineageSampler,
+        },
         coalescence_sampler::unconditional::UnconditionalCoalescenceSampler,
         dispersal_sampler::in_memory::alias::InMemoryAliasDispersalSampler,
         emigration_exit::{domain::DomainEmigrationExit, never::NeverEmigrationExit},
@@ -42,6 +45,155 @@ impl AlgorithmArguments for ClassicalAlgorithm {
     type Arguments = MonolithicArguments;
 }
 
+// Default 'Gillespie' implementation for any turnover sampler
+#[allow(clippy::type_complexity)]
+impl<
+        O: Scenario<
+            IntrinsicsMathsCore,
+            Pcg<IntrinsicsMathsCore>,
+            LineageReference = InMemoryLineageReference,
+        >,
+        R: Reporter,
+        P: LocalPartition<R>,
+    > Algorithm<O, R, P> for ClassicalAlgorithm
+where
+    O::LineageStore<ClassicalLineageStore<IntrinsicsMathsCore, O::Habitat>>:
+        LocallyCoherentLineageStore<IntrinsicsMathsCore, O::Habitat, InMemoryLineageReference>,
+{
+    type Error = !;
+    type LineageReference = InMemoryLineageReference;
+    type LineageStore = O::LineageStore<ClassicalLineageStore<Self::MathsCore, O::Habitat>>;
+    type MathsCore = IntrinsicsMathsCore;
+    type Rng = Pcg<Self::MathsCore>;
+
+    #[allow(clippy::shadow_unrelated, clippy::too_many_lines)]
+    default fn initialise_and_simulate<I: Iterator<Item = u64>>(
+        args: Self::Arguments,
+        rng: Self::Rng,
+        scenario: O,
+        pre_sampler: OriginPreSampler<Self::MathsCore, I>,
+        local_partition: &mut P,
+    ) -> Result<(NonNegativeF64, u64), Self::Error> {
+        match args.parallelism_mode {
+            ParallelismMode::Monolithic => {
+                let lineage_store =
+                    Self::LineageStore::from_origin_sampler(scenario.sample_habitat(pre_sampler));
+                let (habitat, dispersal_sampler, turnover_rate, speciation_probability) =
+                    scenario.build::<InMemoryAliasDispersalSampler<
+                        Self::MathsCore,
+                        O::Habitat,
+                        Self::Rng,
+                    >>();
+                let coalescence_sampler = UnconditionalCoalescenceSampler::default();
+                let emigration_exit = NeverEmigrationExit::default();
+                let event_sampler = UnconditionalEventSampler::default();
+                let immigration_entry = NeverImmigrationEntry::default();
+                let active_lineage_sampler = IndividualAliasActiveLineageSampler::new(
+                    &habitat,
+                    &turnover_rate,
+                    &lineage_store,
+                );
+
+                let simulation = SimulationBuilder {
+                    maths: PhantomData::<Self::MathsCore>,
+                    habitat,
+                    lineage_reference: PhantomData::<Self::LineageReference>,
+                    lineage_store,
+                    dispersal_sampler,
+                    coalescence_sampler,
+                    turnover_rate,
+                    speciation_probability,
+                    emigration_exit,
+                    event_sampler,
+                    active_lineage_sampler,
+                    rng,
+                    immigration_entry,
+                }
+                .build();
+
+                Ok(parallelisation::monolithic::monolithic::simulate(
+                    simulation,
+                    local_partition,
+                ))
+            },
+            non_monolithic_parallelism_mode => {
+                let decomposition =
+                    O::decompose(scenario.habitat(), local_partition.get_partition());
+
+                let rng = rng.split_to_stream(u64::from(local_partition.get_partition().rank()));
+                let lineage_store =
+                    Self::LineageStore::from_origin_sampler(DecompositionOriginSampler::new(
+                        scenario.sample_habitat(pre_sampler),
+                        &decomposition,
+                    ));
+                let (habitat, dispersal_sampler, turnover_rate, speciation_probability) =
+                    scenario.build::<InMemoryAliasDispersalSampler<
+                        Self::MathsCore,
+                        O::Habitat,
+                        Self::Rng,
+                    >>();
+                let coalescence_sampler = UnconditionalCoalescenceSampler::default();
+                let emigration_exit = DomainEmigrationExit::new(decomposition);
+                let event_sampler = UnconditionalEventSampler::default();
+                let immigration_entry = BufferedImmigrationEntry::default();
+                let active_lineage_sampler = IndividualAliasActiveLineageSampler::new(
+                    &habitat,
+                    &turnover_rate,
+                    &lineage_store,
+                );
+
+                let simulation = SimulationBuilder {
+                    maths: PhantomData::<Self::MathsCore>,
+                    habitat,
+                    lineage_reference: PhantomData::<Self::LineageReference>,
+                    lineage_store,
+                    dispersal_sampler,
+                    coalescence_sampler,
+                    turnover_rate,
+                    speciation_probability,
+                    emigration_exit,
+                    event_sampler,
+                    active_lineage_sampler,
+                    rng,
+                    immigration_entry,
+                }
+                .build();
+
+                match non_monolithic_parallelism_mode {
+                    ParallelismMode::Monolithic => unsafe { unreachable_unchecked() },
+                    ParallelismMode::Optimistic(OptimisticParallelismMode { delta_sync }) => {
+                        Ok(parallelisation::monolithic::optimistic::simulate(
+                            simulation,
+                            delta_sync,
+                            local_partition,
+                        ))
+                    },
+                    ParallelismMode::Lockstep => {
+                        Ok(parallelisation::monolithic::lockstep::simulate(
+                            simulation,
+                            local_partition,
+                        ))
+                    },
+                    ParallelismMode::OptimisticLockstep => {
+                        Ok(parallelisation::monolithic::optimistic_lockstep::simulate(
+                            simulation,
+                            local_partition,
+                        ))
+                    },
+                    ParallelismMode::Averaging(AveragingParallelismMode { delta_sync }) => {
+                        Ok(parallelisation::monolithic::averaging::simulate(
+                            simulation,
+                            delta_sync,
+                            local_partition,
+                        ))
+                    },
+                }
+            },
+        }
+    }
+}
+
+// Optimised 'Classical' implementation for the `UniformTurnoverSampler`
 #[allow(clippy::type_complexity)]
 impl<
         O: Scenario<
@@ -57,12 +209,6 @@ where
     O::LineageStore<ClassicalLineageStore<IntrinsicsMathsCore, O::Habitat>>:
         LocallyCoherentLineageStore<IntrinsicsMathsCore, O::Habitat, InMemoryLineageReference>,
 {
-    type Error = !;
-    type LineageReference = InMemoryLineageReference;
-    type LineageStore = O::LineageStore<ClassicalLineageStore<Self::MathsCore, O::Habitat>>;
-    type MathsCore = IntrinsicsMathsCore;
-    type Rng = Pcg<Self::MathsCore>;
-
     #[allow(clippy::too_many_lines)]
     fn initialise_and_simulate<I: Iterator<Item = u64>>(
         args: Self::Arguments,
