--- rustcoalescence/src/simulation/mod.rs
@@ -1,4 +1,0 @@
-pub mod almost_infinite;
-pub mod spatially_explicit;
-pub mod non_spatial;
-pub mod spatially_implicit;
